Add files via upload
The big difference between objects and arrays is that in objects order of these values does not matter at all when we want to retrieve them. But in arrays the order of elements in which we specify matters the most. It means that object is basically used when the data is unstructured. //But in the case whenever u do not know that which property i want to extract meaning i want the user to enter the property manually randomly. Then in that case we use prompt function and bracket notation.Dot notation does not works in this case.